@@ -1,8 +1,9 @@
 package com.github.razir.progressbutton
 
-import android.support.annotation.ColorInt
-import android.support.annotation.ColorRes
-import android.support.annotation.DimenRes
+import androidx.annotation.ColorInt
+import androidx.annotation.ColorRes
+import androidx.annotation.DimenRes
+
 
 /**
  * Use to customize your progress drawable
