@@ -1,13 +1,13 @@
 package com.github.razir.progressbutton
 
 import android.animation.Animator
-import android.arch.lifecycle.Lifecycle
-import android.arch.lifecycle.LifecycleObserver
-import android.arch.lifecycle.LifecycleOwner
-import android.arch.lifecycle.OnLifecycleEvent
 import android.graphics.drawable.Animatable
 import android.view.View
 import android.widget.TextView
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleObserver
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.OnLifecycleEvent
 import java.lang.ref.WeakReference
 import java.util.*
 
@@ -37,7 +37,8 @@ fun TextView.cleanUpDrawable() {
     }
 }
 
-private class ProgressButtonHolder(private val textView: WeakReference<TextView>) : LifecycleObserver {
+private class ProgressButtonHolder(private val textView: WeakReference<TextView>) :
+    LifecycleObserver {
 
     @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
     fun onDestroy() {
